Renamed file that did not match the class name
