+++ src/contentProcessing/pointClouds/GltfTransformPointClouds.ts
